pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atomization_scorer"
version = "0.1.0"

description = """
Analytical tool to evaluate the quality and consistency of genome atomization results,
providing quantitative and visual assessments.
"""

authors = [
  { name = "Matej Krivosik", email = "krivosik7@uniba.sk" }
]

readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-only" }

keywords = [
  "genomics",
  "atomization",
  "IMP algorithm",
  "GEESE",
  "bioinformatics",
  "sequence analysis",
]

dependencies = [
  "pandas>=2.2,<3",
  "biopython>=1.83,<2",
  "intervaltree>=3.1,<4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<10",
  "ruff>=0.6,<1",
  "mypy>=1.10,<2",
]

[project.scripts]
atomization_scorer = "atomization_scorer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/atomization_scorer"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
